[feature/RSCRUD-5] - Updated POST body modifier for work with promises
