+++ Node_CRUD/src/server/utils/MiddlewarePipeline/MiddlewarePipeline.ts
@@ -1,0 +1,8 @@
+// Middleware explanation
+// https://muniftanjim.dev/blog/basic-middleware-pattern-in-javascript/
+
+import { TMiddleware } from './types';
+
+export class MiddlewarePipeline<TReq, TRes> {
+  constructor(...middlewares: TMiddleware<TReq, TRes>[]) {}
+}
