+++ Node_CRUD/src/server/utils/MiddlewarePipeline/types.ts
@@ -1,0 +1,7 @@
+export type TNext = () => Promise<void> | void;
+
+export type TMiddleware<TRequest, TResponse> = (
+  req: TRequest,
+  res: TResponse,
+  next: TNext
+) => Promise<void> | void;
